[#28] Moved buttons in a separate view
