+++ AnimalPocket/AnimalPocket/Views/BugItems/BugDetails/BugDetails.swift
