+++ AnimalPocket/AnimalPocket/Views/BugItems/BugDetails/BugImage.swift
